refactor(companion): extract BookingDetailScreen with AppPressable migration
